+++ src/main/java/com/mraof/minestuck/client/ClientProxy.java
@@ -8,7 +8,7 @@ import com.mraof.minestuck.client.model.armor.*;
 import com.mraof.minestuck.client.model.entity.BishopModel;
 import com.mraof.minestuck.client.model.entity.RookModel;
 import com.mraof.minestuck.client.model.MSModelLayers;
-import com.mraof.minestuck.particles.TransportalizerParticle;
+import com.mraof.minestuck.client.particles.TransportalizerParticle;
 import com.mraof.minestuck.client.renderer.blockentity.*;
 import com.mraof.minestuck.client.renderer.entity.*;
 import com.mraof.minestuck.client.renderer.entity.frog.FrogRenderer;
@@ -19,7 +19,7 @@ import com.mraof.minestuck.item.BoondollarsItem;
 import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.item.block.StoneTabletItem;
 import com.mraof.minestuck.item.weapon.MusicPlayerWeapon;
-import com.mraof.minestuck.particles.MSParticleType;
+import com.mraof.minestuck.util.MSParticleType;
 import com.mraof.minestuck.world.MSDimensions;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.HumanoidModel;
